Include support for GATK license key.
The license key is a binary file that is base64'd and then stored
in a variable under ngi_pipeline/vars/. This var gets unencrypted
by the password stored in .vault_password.txt, and piped to a
temporary file.

As this temporary file is in base64 we have to convert it back to
binary, before we can deploy it to its destination and then
remove the temporary files.