Cleaning up under /lupus/ngi on irma3.
Moved the playbooks to a deploy folder, and
renamed the playbook to install.yml to make it a bit more clear that the deploy
process is a two-step rocket that involves installing sw with Ansible and then
syncing the sw with the sync.py script.

Updated the README with a bit more info about how the environment should be
bootstrapped.